requirement.txt
streamlit==1.29.0
groq==0.10.0
Pillow==10.0.0